Continue refactoring and adding deploy logic for SDK generation.
