+++ src/generateSdk/sdkWriter/sdkWriter.ts
@@ -2,25 +2,30 @@
 import { SdkGeneratorResponse } from '../../models/sdkGeneratorResponse.js';
 import { ClassUrlMap } from '../../utils/sdk.js';
 import { Language } from '../../yamlProjectConfiguration/models.js';
-import { YamlFrontend } from '../../yamlProjectConfiguration/v1.js';
 import { basicFileWriter } from './basicFileWriter.js';
 import { writeSdkTs, writeSdkJs } from './jsSdkWriter.js';
 
-export async function writeSdk(language: Language, frontend: YamlFrontend|undefined, projectName: string, projectRegion: string, stage: string, sdkResponse: SdkGeneratorResponse, classUrls: ClassUrlMap[], publish: boolean, path: string|undefined) {
+export async function writeSdk(
+    language: Language, 
+    projectName: string, 
+    projectRegion: string,
+    stage: string,
+    sdkResponse: SdkGeneratorResponse,
+    classUrls: ClassUrlMap[],
+    publish: boolean, 
+    outputPath: string|undefined): Promise<string> {
     switch (language) {
         case Language.ts:
-            writeSdkTs(projectName, frontend, projectRegion, stage, sdkResponse, classUrls, publish);
-            break;
+            return await writeSdkTs(projectName, projectRegion, stage, sdkResponse, classUrls, publish, outputPath);
         case Language.js:
-            writeSdkJs(projectName, frontend, projectRegion, stage, sdkResponse, classUrls, publish);
-            break;
+            return await writeSdkJs(projectName, projectRegion, stage, sdkResponse, classUrls, publish, outputPath);
         case Language.go:
         case Language.kt:
         case Language.dart:
         case Language.swift:
         case Language.python:
-            basicFileWriter(sdkResponse, path!);
-            break;
-
+            return await basicFileWriter(sdkResponse, outputPath!);
+        default:
+            throw new Error(`Language ${language} is not supported`);
     }
 }
