+++ src/generateSdk/utils/compileSdk.ts
@@ -94,6 +94,7 @@ export async function compileSdk(
     workers.push(writePackagePromise);
     await Promise.all(workers);
     if (publish === true) {
+        console.log("Publishing the SDK...", modulePath);
         await packageManager.publish(modulePath);
     }
 }
